refactor: fixed components and rename files
