+++ src/components/tables/ReportsTable.vue
@@ -1,0 +1,10 @@
+<script setup>
+import { ref } from 'vue';
+</script>
+
+<template>
+</template>
+
+<style scoped>
+</style>
+
