+++ src/pages/user/UsersManagement.vue
@@ -1,0 +1,21 @@
+<script setup>
+import UsersTabNavigation from '../../components/tabs/UsersTabNavigation.vue';
+</script>
+
+<template>
+    <section class="w-full min-h-screen p-4">
+        <!-- Title -->
+        <div class="mt-6 px-2 flex justify-between items-center">
+            <h1 class="text-2xl font-bold dark:text-white text-center sm:text-left">
+                Users Management
+            </h1>
+        </div>
+        
+        <!-- Tab Navigation -->
+        <UsersTabNavigation />
+    </section>
+</template>
+
+<style scoped>
+</style>
+
